Reset some bigger changes by relocating data method
